update and tested >=torch1.9 FFT and RMSD code

+++ Models/BruteForce/validation_metrics.py
@@ -47,26 +47,26 @@ class RMSD:
         T2 = translation2.clone().detach().cuda()
         T = T1 - T2
 
-        rotation1 = torch.tensor([rotation1], dtype=torch.float).cuda()
-        rotation2 = torch.tensor([rotation2], dtype=torch.float).cuda()
-        R1 = torch.zeros(2, 2, dtype=torch.float).cuda()
+        rotation1 = torch.tensor([rotation1], dtype=torch.float64).cuda()
+        rotation2 = torch.tensor([rotation2], dtype=torch.float64).cuda()
+        R1 = torch.zeros(2, 2, dtype=torch.float64).cuda()
         R1[0, 0] = torch.cos(rotation1)
         R1[1, 1] = torch.cos(rotation1)
         R1[1, 0] = torch.sin(rotation1)
         R1[0, 1] = -torch.sin(rotation1)
-        R2 = torch.zeros(2, 2, dtype=torch.float).cuda()
+        R2 = torch.zeros(2, 2, dtype=torch.float64).cuda()
         R2[0, 0] = torch.cos(rotation2)
         R2[1, 1] = torch.cos(rotation2)
         R2[1, 0] = torch.sin(rotation2)
         R2[0, 1] = -torch.sin(rotation2)
         R = R2.transpose(0, 1) @ R1
 
-        I = torch.diag(torch.ones(2, dtype=torch.float)).cuda()
+        I = torch.diag(torch.ones(2, dtype=torch.float64)).cuda()
         # RMSD
         rmsd = torch.sum(T * T)
         rmsd = rmsd + torch.sum((I - R) * X, dim=(0, 1))
         rmsd = rmsd + 2.0 * torch.sum(torch.sum(T.unsqueeze(dim=1) * (R1 - R2), dim=0) * C, dim=0) + self.epsilon
-
+        # print(rmsd)
         return torch.sqrt(rmsd)
 
     def calc_rmsd(self):
